# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shipbox API Documentation",
        "description": "API Endpoints for Shipbox",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-post": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Creates an order for the authenticated user or for a specific user if the requester is an admin. Validates that all products belong to the order owner.",
                "operationId": "afa45dec6d3ec58823e62ea881a5878e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_phone",
                                    "customer_address",
                                    "products",
                                    "total_price"
                                ],
                                "properties": {
                                    "customer_phone": {
                                        "description": "Max 20 characters",
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "123 Main St, City"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Please deliver after 5 PM",
                                        "nullable": true
                                    },
                                    "total_price": {
                                        "description": "Cumulative price of all products",
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "user_id": {
                                        "description": "Required only if requester is an admin. Specifies the owner of the order.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "description": "The product ID",
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "order": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order_id": {
                                                    "type": "string",
                                                    "example": "ORD-ABC12345"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "customer_phone": {
                                                    "type": "string",
                                                    "example": "0123456789"
                                                },
                                                "customer_address": {
                                                    "type": "string",
                                                    "example": "123 Main St, City"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "to prepare"
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150.5
                                                },
                                                "quantity": {
                                                    "description": "Total quantity of all items",
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 101
                                                            },
                                                            "pivot": {
                                                                "properties": {
                                                                    "order_id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "product_id": {
                                                                        "type": "integer",
                                                                        "example": 101
                                                                    },
                                                                    "quantity": {
                                                                        "type": "integer",
                                                                        "example": 2
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Attempting to order products not owned by the target user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden. You can only create orders for products you own."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The customer phone field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create order"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/orders/{order}/delete": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "description": "Permanently deletes an order and removes its associations with products. Sellers can delete their own orders, while Admins can delete any order.",
                "operationId": "f450b501d11677da9e42c2c5fcb8a6f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Unauthorized to delete this order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized to delete this order."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Order] 123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}